added new autonomous (hubpark) not done/needs testing tried to fix setting the linear slide with one button a few ease of mind changes

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/FieldMeasurements.java
@@ -5,20 +5,20 @@ public class FieldMeasurements {
     public static final double TILE_WIDTH = 610; //3728.44037 in encoder ticks //2.15 per two tiles
     //1.0.75 for 1 tile  at speed 0.5
     public static final long TIME_FOR_TILE = 1075; //in MS
-    private static final double STAGEONE_HEIGHT = -3300;    //IN TICKS
-    private static final double STAGETWO_HEIGHT = -5300;   //IN TICKS
-    private static final double STAGETHREE_HEIGHT = -6700;
+    private static final int STAGEONE_HEIGHT = -3300;    //IN TICKS
+    private static final int STAGETWO_HEIGHT = -5300;   //IN TICKS
+    private static final int STAGETHREE_HEIGHT = -6700;
 
     public static int getStageOneHeight() {
-        return (int)STAGEONE_HEIGHT;
+        return STAGEONE_HEIGHT;
     }
 
     public static int getStageTwoHeight() {
-        return (int)STAGETWO_HEIGHT;
+        return STAGETWO_HEIGHT;
     }
 
     public static int getStageThreeHeight() {
-        return (int)STAGETHREE_HEIGHT;
+        return STAGETHREE_HEIGHT;
     }
 
 }
